# angular-app/src/openapi/openapi3.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Group Service API",
    "description": "API for managing educational groups secured with Keycloak",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8083",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/notifications/{notificationId}/read": {
      "put": {
        "tags": [
          "notification-controller"
        ],
        "operationId": "markAsRead",
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/notifications/{notificationId}": {
      "delete": {
        "tags": [
          "notification-controller"
        ],
        "operationId": "deleteNotification",
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Provide the JWT token obtained from Keycloak: 'Bearer eyJ...'",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}